add check docker group

+++ scripts/down.sh
@@ -1,3 +1,12 @@
 #!/bin/sh
-cd $(dirname $0)/.. || exit 1
-docker-compose down
+gd=docker
+gr=$(groups "$USER" | grep -o -w "\b$gd\b")
+
+if [ "$gd" = "$gr" ]; then
+    cd $(dirname $0)/.. || exit 1
+    docker-compose down
+    exit
+else
+    echo "User $USER is not a member of the Docker group."
+    exit 1
+fi
